courselist/requirements.txt
boto3==1.9.115
botocore==1.12.253
Django==2.1.5
django-cors-headers==3.2.1
django-environ==0.4.5
django-getenv==1.3.2
djangorestframework==3.7.3
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
s3transfer==0.2.1
six==1.14.0
urllib3==1.25.8
